requirements.txt
pdf2image==1.16.0
Pillow==9.2.0
pytesseract==0.3.10
SpeechRecognition==3.8.1
